[refactor/create_module-client] 메일건서비스를 이용하는 mailgun-client 생성
메일링을 위한 모듈로 분리
실행가능한 userApplication 역시 openfeign 의존성이 필요함
bean 을
@EnableFeignClients(basePackages = "com.zerobase.client") 추가
