+++ Community/build.gradle
@@ -25,11 +25,13 @@ subprojects {
     apply plugin : "io.spring.dependency-management"
     dependencies {
 
-        //database
-        runtimeOnly 'com.h2database:h2'
-        //jpa
-        implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-        //lombok
+        if (project != project(':module-client')) {
+            //database
+            runtimeOnly 'com.h2database:h2'
+            //jpa
+            implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+            //lombok
+        }
         compileOnly 'org.projectlombok:lombok'
         annotationProcessor 'org.projectlombok:lombok'
         //test
@@ -52,5 +54,6 @@ project(':module-admin') {
 project(':module-user') {
     dependencies {
         implementation project(':module-common')
+        implementation project(':module-client')
     }
 }
