+++ Community/module-admin/README.md
@@ -1,0 +1,5 @@
+## module-admin
+---
+
+- 백오피스를 담당합니다.
+- 실행가능한 application 으로 admin service application 입니다.
