+++ Community/module-client/README.md
@@ -1,0 +1,5 @@
+### module-client
+---
+
+- 네트워크로 연결된 서버로부터 정보를 제공받는 모듈입니다.
+- 메일링을 담당하고 있습니다.
