+++ Community/module-client/src/main/java/com/zerobase/client/mailgun/config/FeignConfig.java
@@ -1,11 +1,10 @@
-package com.zerobase.user.config;
+package com.zerobase.client.mailgun.config;
 
 import feign.auth.BasicAuthRequestInterceptor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
 @Configuration
 public class FeignConfig {
 	@Value(value = "${mailgun.key}")
