+++ Community/module-common/README.md
@@ -1,0 +1,6 @@
+## module-common
+---
+
+- 코드의 재활용성을 위해 자주쓰이는 코드들의 집합을 담당합니다.
+- entity 와 repository 를 담당합니다.
+- 비대해지지않도록 주의해야합니다.
