+++ Community/module-user/README.md
@@ -1,0 +1,6 @@
+## module-user
+---
+
+- member 의 authentication 을 담당합니다. 
+- member 의 registration 을 담당합니다.
+- 실행가능한 application 으로 service application 입니다.
