+++ Community/module-user/build.gradle
@@ -12,7 +12,6 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.apache.commons:commons-lang3:3.12.0'
     implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
-
 }
 dependencyManagement {
     imports {
