+++ Community/module-user/src/main/java/com/zerobase/user/UserApplication.java
@@ -3,18 +3,17 @@ package com.zerobase.user;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.boot.web.servlet.ServletComponentScan;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = "com.zerobase")
 @EntityScan(basePackages = "com.zerobase.common")
 @EnableJpaRepositories(basePackages = "com.zerobase.common")
 @EnableJpaAuditing
-@EnableFeignClients
-@ServletComponentScan
+@EnableFeignClients(basePackages = "com.zerobase.client")
+//@ServletComponentScan
 public class UserApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(UserApplication.class, args);
